test: characterize atlas idle appended lineage
