+++ src/hooks/atlas/idle-constants.ts
@@ -1,0 +1,4 @@
+export const CONTINUATION_COOLDOWN_MS = 5000
+export const FAILURE_BACKOFF_MS = 5 * 60 * 1000
+export const MAX_CONSECUTIVE_PROMPT_FAILURES = 10
+export const RETRY_DELAY_MS = CONTINUATION_COOLDOWN_MS + 1000
